refactor(tools): split generate_alertmanager_routes.py into 5 helper modules (PR-4 of 4)
generate_alertmanager_routes.py was 1645 LOC, the largest file under
scripts/tools/ops/ and a clear god-file. Split into 5 helper modules
along the section boundaries the original author had already drawn,
plus a thin facade that re-exports every public + private symbol so
existing test imports (25+ symbols across 9 test files) work unchanged.

Module layout:

  _grar_validate.py (240 LOC) — URL/domain/schema validation
  _grar_merge.py    (130 LOC) — routing merge + tenant substitution
  _grar_parse.py    (296 LOC) — YAML config loading + 4-layer merge
  _grar_routes.py   (556 LOC) — route generation + override expansion + inhibit
  _grar_render.py   (307 LOC) — output rendering + ConfigMap operations

  generate_alertmanager_routes.py (1645 → 309 LOC, 81% reduction)
    Re-exports + 5 CLI mode handlers + main() + argparse.

Net LOC: 1645 → 1838 across 6 files (+12% overhead from per-module
docstrings / imports), but no single file > 600 LOC anymore.

Verification:
  * 322/322 tests pass across the 9 test files that import from
    generate_alertmanager_routes (full list in commit body).
  * No public or private symbol changes — every test import works
    through the facade re-export pattern.
  * build.sh TOOL_FILES + _lint_helpers BUILD_EXEMPT updated so the
    5 helpers ship in the docker image and are exempt from CLI-coverage
    checks. tool-map.md / tool-map.en.md regenerated (140 → 146 tools).
  * .sed-damage-allowlist entry added for the 80% shrink (genuine
    refactor, not sed -i damage).

This is PR-4 of 4 (the final PR in the v2.8.0 da-tools refactor sweep).
A future cleanup PR can apply the same split pattern to the other
1000+ LOC ops/*.py files (init_project, scaffold_tenant, migrate_rule,
operator_generate, migrate_to_operator).
