+++ scripts/tools/lint/_lint_helpers.py
@@ -28,6 +28,14 @@ BUILD_EXEMPT = frozenset({
     # v2.8.0 PR-2 — shared dispatcher absorbs ~95% of guard /
     # batchpr / parser dispatcher boilerplate. Library, not CLI.
     "_lib_godispatch.py",
+    # v2.8.0 PR-3a — generate_alertmanager_routes.py split into 5 helpers.
+    # These are library modules consumed by the main file via re-export,
+    # not CLI commands themselves.
+    "_grar_validate.py",
+    "_grar_merge.py",
+    "_grar_parse.py",
+    "_grar_routes.py",
+    "_grar_render.py",
     "metric-dictionary.yaml",
     "generate_tenant_mapping_rules.py",
     # v2.8.0 Phase B Track A A5: ship-but-not-public CLI design tradeoff.
